fix(option picker ref incorrectly assigned): this fixes sort filters to be applied correctly
previously the ref is assigned so that options can be resetted and that the menu will be scrolled to
top after submission of options. However, this broke the existing sort filters assignment function.
While ref prop is now correctly given, it is unclear on how to reset the options via onMenuLeave as
those events didn't seem to be captured - this will be a future TODO